Introduce lightweight WorkQueue and KeyedLock
Summary:
- CapConcurrency is very heavyweight and not easily layered.
- LockManager is very heavyweight and ends up consuming a thread
  anytime something needs to block.
  This ends up creating _tons_ of extra threads that are just waiting
  on a semaphore.
- Bring out a very simple "WorkQueue" and then layer a "KeyedLock" on it
  that provides all of this behavior in a non-blocking manner.

A non-trivial amount of time is spent in CapConcurrency and LockManager when looking at performance data.
In addition, creating a group with a TON of apps will create thousands of threads and will likely bring
down the JVM as a result. This should avoid creating a large percentage of the threads.

Test Plan: sbt test

Reviewers: jdef, aquamatthias, mpark

Subscribers: marathon-team

JIRA Issues: DCOS-9121, MARATHON-1147, MARATHON-1108

Differential Revision: https://phabricator.mesosphere.com/D171