escape leader-latch lock
Summary:
Brad Peters in https://github.com/mesosphere/marathon/pull/4640 points
out that LeaderLatch.setLeadership has a lock, and that this can lead
to deadlock situations.

In this patch we change the approach such that leadership change
events are sent asynchronously to CuratorElectionService, freeing the
LeaderLatch setLeadership lock to be closed while we try to obtain
others in a separate thread.

Test Plan: test leader election code; collaborate with Brad Peters to see if it helps

Reviewers: jasongilanfarr, jdef

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D377