Use 1 master and 1 agent to fix #4851.
Summary:
The test was flaky because the app was deployed on one of two
agents. When it then killed one agent it was not sure if it killed the
right one.

We also ensure that Marathon suicides and does not lockup during shutdown.
This diff depends on D379.

Test Plan: integration-test

Reviewers: timcharper, jdef, zen-dog

Subscribers: marathon-team, jenkins

Differential Revision: https://phabricator.mesosphere.com/D391